Fix error the $ZSH folder already exists
Running the core states a second time gives the error the $ZSH folder
already exists (/home/dylan/.oh-my-zsh) and causes the minion execution
to hang (likely because it drops back into a Zsh session).

This is caused by the creates argument using ~ which refers to /root
instead of /home/dylan.  Replaced with an absolute value.